Implement further changes to the RHF API.
