+++ gqcp/include/QCMethod/RHF/RHF.hpp
@@ -63,65 +63,4 @@ public:
 };
 
 
-/*
- *  HELPER METHODS
- */
-
-/**
- *  Calculate the RHF Fock matrix F = H_core + G, in which G is a contraction of the density matrix and the two-electron integrals
- *
- *  @param D_AO                 the RHF density matrix in AO basis
- *  @param sq_hamiltonian       the Hamiltonian expressed in an AO basis
- *
- *  @return the RHF Fock matrix expressed in the AO basis
- */
-ScalarSQOneElectronOperator<double> calculateRHFAOFockMatrix(const OneRDM<double>& D_AO, const SQHamiltonian<double>& sq_hamiltonian);
-
-/**
- *  @param D_AO         the RHF density matrix in AO basis
- *  @param H_core_AO    the core Hamiltonian expressed in an AO basis
- *  @param F_AO         the Fock matrix in AO basis
- *
- *  @return the RHF electronic energy
- */
-double calculateRHFElectronicEnergy(const OneRDM<double>& D_AO, const ScalarSQOneElectronOperator<double>& H_core_AO, const ScalarSQOneElectronOperator<double>& F_AO);
-
-/**
- *  @param N    the number of electrons
- *
- *  @return the RHF HOMO index
- */
-size_t RHFHOMOIndex(size_t N);
-
-/**
- *  @param K    the number of spatial orbitals
- *  @param N    the number of electrons
- *
- *  @return the RHF LUMO index
- */
-size_t RHFLUMOIndex(size_t K, size_t N);
-
-/**
- *  Specialize the orbital Hessian for RHF
- * 
- *  @param sq_hamiltonian       the Hamiltonian expressed in an orthonormal basis
- *  @param N_P                  the number of electron pairs
- * 
- *  @return the RHF orbital Hessian as a BlockRankFourTensor, i.e. an object with a suitable operator() implemented
- */
-BlockRankFourTensor<double> calculateRHFOrbitalHessianTensor(const SQHamiltonian<double>& sq_hamiltonian, const size_t N_P);
-
-/**
- *  @param sq_hamiltonian       the Hamiltonian expressed in an orthonormal basis
- *  @param N_P                  the number of electron pairs
- *  @param a                    the first virtual orbital index
- *  @param i                    the first occupied orbital index
- *  @param b                    the second virtual orbital index
- *  @param j                    the second occupied orbital index
- * 
- *  @return an element of the RHF orbital Hessian
- */
-double calculateRHFOrbitalHessianElement(const SQHamiltonian<double>& sq_hamiltonian, const size_t N_P, const size_t a, const size_t i, const size_t b, const size_t j);
-
-
 }  // namespace GQCP
