+++ gqcp/include/QCModel/RHF/CMakeLists.txt
@@ -1,0 +1,4 @@
+target_sources(gqcp
+    PRIVATE
+        RHF.hpp
+)
