+++ gqcp/src/QCMethod/RHF/RHFSCFSolver.cpp
@@ -17,7 +17,7 @@
 // 
 #include "QCMethod/RHF/RHFSCFSolver.hpp"
 
-#include "QCModel/RHF.hpp"
+#include "QCModel/RHF/RHF.hpp"
 
 
 namespace GQCP {
@@ -108,7 +108,7 @@ void RHFSCFSolver::solve(const TransformationMatrix<double>& C_initial) {
             }
 
             // Set the converged solution
-            auto electronic_energy = calculateRHFElectronicEnergy(D_AO, H_core, F_AO);
+            auto electronic_energy = QCModel::RHF<double>::calculateElectronicEnergy(D_AO, H_core, F_AO);
             this->solution = RHF(electronic_energy, TransformationMatrix<double>(C), generalized_eigensolver.eigenvalues());
 
         } else {  // not converged yet
