+++ gqcp/tests/QCMethod/RHF/CMakeLists.txt
@@ -1,7 +1,6 @@
 list(APPEND test_target_sources
     ${CMAKE_CURRENT_SOURCE_DIR}/DIISRHFSCFSolver_test.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/PlainRHFSCFSolver_test.cpp
-    ${CMAKE_CURRENT_SOURCE_DIR}/QCMethod_RHF_test.cpp
 )
 
 set(test_target_sources ${test_target_sources} PARENT_SCOPE)
